% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSFDR.R
\name{optim.arbidept.adjp}
\alias{optim.arbidept.adjp}
\title{Objective function to be optimized (4)}
\usage{
optim.arbidept.adjp(alpha, p, k)
}
\arguments{
\item{alpha}{the parameter we need to solve for the adjusted p-values.}

\item{p}{numeric vector of p-values (possibly with \code{\link[base]{NA}}s). Any other R is coerced by \code{\link[base]{as.numeric}}. Same as in \code{\link[stats]{p.adjust}}.}

\item{k}{pre-specified number of acceptances allowed in the testing procedure (cannot exceed the length of \code{p})}
}
\value{
difference between adjusted p-value and significant level alpha.
}
\description{
Objective function to be optimized for the adjusted p-values for FDR controlling g-FSMTP under arbitrary dependence. (See Theorem 3.1 and Theorem 4.1 in Lynch et al. (2016))
}
\author{
Yalin Zhu
}
\references{
Lynch, G., Guo, W., Sarkar, S. K., & Finner, H. (2016).
 The Control of the False Discovery Rate in Fixed Sequence Multiple Testing.
 \emph{arXiv preprint} arXiv:1611.03146.
}

